no-op balance update in case of match value in db, fix types to match BIGINT

+++ badges/achieve_badges_test.go
@@ -232,8 +232,8 @@ func defaultCfg() *config {
 func badgeProgress(alreadyAchieved *users.Enum[Type], balance, friends, levels uint64) *progress {
 	return &progress{
 		AchievedBadges:  alreadyAchieved,
-		Balance:         balance,
-		FriendsInvited:  friends,
-		CompletedLevels: levels,
+		Balance:         int64(balance),
+		FriendsInvited:  int64(friends),
+		CompletedLevels: int64(levels),
 	}
 }
